move preprocessor definitions to arrow/config.h.cmake
