+++ cpp/CMakeLists.txt
@@ -135,7 +135,6 @@ endif()
 # ----------------------------------------------------------------------
 # cmake options
 include(DefineOptions)
-add_definitions_for_options()
 
 # Needed for linting targets, etc.
 if(${CMAKE_VERSION} VERSION_LESS "3.12.0")
@@ -825,3 +824,10 @@ config_summary_message()
 if(${ARROW_BUILD_CONFIG_SUMMARY_JSON})
   config_summary_json()
 endif()
+
+# Write out arrow and third party package configurations.
+
+configure_file("src/arrow/config.h.cmake" "src/arrow/config.h")
+
+install(FILES "${ARROW_BINARY_DIR}/src/arrow/config.h"
+        DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/arrow")
